refactor: extract _conv_block to models/blocks.py, fix gradient test
